Added Photino Logo SVG
The logo was a rough idea built with CSS and was not very
portable or scalable.

I was able to replace it with a 100% SVG logo that even
contains the animation in the SVG itself instead of having
to use CSS animations for it. It will likely be easier to create
other image formats like png, gif and ico versions.