Fixed quick start commands
The `cd MyProjects` command tries to open a folder that
is not created beforehand and breaks the quick start.
It was removed because the `new` command creates a
folder with the `-o` option. The subsequent `cd` into
that folder is correct.

The command to list templates should remain informative
in the comments. It is not interactive and the terminal
would simply scroll past the list. Especially if the
user is using the copy function of the terminal on the
bottom right hand corner of the container.